Obtain training data, testing data, and train signature's original model.

Obtain training data, testing data, and train signature's original model.

ObtainSampleScore_OriginalModel(
  theObject_train,
  useAssay,
  gene_set,
  input,
  SigName,
  obtainDiagnosis,
  annotationColName,
  FUN,
  adj
)

Arguments

theObject_train
A SummarizedExperiment object that has been prestored in the data file: OriginalTrainingData.

useAssay
A character string or an integer specifying the assay in the input. Used for the test SummarizedExperiment object. Default is 1, indicating the first assay in the test SummarizedExperiment object.

gene_set
A character vector that includes gene symbols for selected gene signature.

input
A SummarizedExperiment object with gene symbols as the assay row names.

SigName
Optional. A character string that indicates the name for gene_set. SigName is used to provide information when gene signatures were missing in the test data.

obtainDiagnosis
Boolean. Used to create training data if TRUE. Default is FALSE

annotationColName
A character string specifying the column name of disease status. Only used when creating training data. Default is NULL.

FUN
A character string specifying the function name of the corresponding signature's original model.

adj
A small real number used in combat to solve for genes with 0 counts in rare cases. Not required for most of cases.

Value

The predicted score for each sample in the test study using corresponding gene signature's original model.