TB gene signatures that require retraining.

A function to obtain predicted score for TB gene signatures that need retraining of original models.

Usage

.OriginalModel_Retraining(input, useAssay, geneSignaturesName, adj, BPPARAM)

Arguments

input

A SummarizedExperiment object with gene symbols as the assay row names.

useAssay

A character string or an integer specifying the assay in the input. Default is the first assay in the assay list. Used for the test SummarizedExperiment object. Default is 1, indicating the first assay in the input.

geneSignaturesName

A character string/vector specifying the signature of interest. If any(geneSignaturesName == "") == TRUE, it will run all available gene signatures' original models.

adj

A small positive real number used in ComBat to solve for genes with 0 counts(rare cases). Default is 1e-3.

BPPARAM

An instance inherited from bplapply.

Value

A SummarizedExperiment object with predicted scores for each sample obtained from the signature's original model.

Details

Maertzdorf_4 and Maertzdorf_15 were trained using a random forest to distinguish patients with active TB from healthy controls.

Verhagen_10 was also trained using a random forest to distinguish samples with active TB from either latent infection or healthy controls. The random forest model was build using randomForest.

Jacobsen_3 were trained using linear discriminant analysis (LDA) to distinguish samples with active TB from latent infection status.

Sambarey_HIV_10 were also trained using LDA to distinguish samples with active TB from either latent infection, healthy control, or other disease (HIV). The LDA model was built using lda.

Berry_OD_86 and Berry_393 were trained using K-nearest neighbors (KNN) model to differentiate samples with active TB from latent infection status. The KNN model was built using knn.

Suliman_RISK_4 and Zak_RISK_16 were trained using support vector machines (SVM) to distinguish TB progressor from non-progressors. The input gene expression features for Suliman_RISK_4 used the paired ratio of GAS6/CD1C, SEPTIN4/BLK, SEPTIN4/CD1C, GAS6/BLK. The SVM model was built using svm.