A function that implements the original models for various TB signatures.

TB signatures with available original models are: Anderson_42, Anderson_OD_51, Kaforou_27, Kaforou_OD_44, Kaforou_OD_53, Sweeney_OD_3, Maertzdorf_4, Verhagen_10, Jacobsen_3, Sambarey_HIV_10, Leong_24, Berry_OD_86, Berry_393, Bloom_OD_144, Suliman_RISK_4, Zak_RISK_16, Leong_RISK_29, and Zhao_NANO_6. The predicted score for each signature has been stored in the column data section of the input SummarizedExperiment study.

evaluateOriginalModel(
  input,
  useAssay = 1,
  geneSignaturesName = NULL,
  adj = 0.001,
  BPPARAM = BiocParallel::SerialParam(progressbar = TRUE)
)

Arguments

input
A SummarizedExperiment object with gene symbols as the assay row names.

useAssay
A character string or an integer specifying the assay in the input. Used for the test SummarizedExperiment object. Default is 1, indicating the first assay in the input.

geneSignaturesName
A character string/vector specifying the signature of interest. Default is NULL. If geneSignaturesName is NULL, run all available gene signatures' original model in the package.

adj
A small real number used in combat to solve for genes with 0 counts in rare cases.

BPPARAM
An instance inherited from bplappy. See bplapply for details.

Value

A SummarizedExperiment object with predicted scores for each sample obtained from the signature's original model.

Examples

data("TB_hiv", package = "TBSignatureProfiler")
re <- evaluateOriginalModel(input = TB_hiv, useAssay = "counts",
geneSignaturesName = c("Anderson_42", "Maertzdorf_15"))

#>   |                                                                              |                                                                      |   0%

#> Now Evaluating: Anderson_42

#>   |                                                                              |======================================================================| 100%
#> 
#>   |                                                                              |                                                                      |   0%

#> Now Evaluating: Maertzdorf_15

#> Using batch =GSE_trainas a reference batch (this batch won't change)

#> Found2batches

#> Adjusting for0covariate(s) or covariate level(s)

#> Standardizing Data across genes

#> Fitting L/S model and finding priors

#> Finding parametric adjustments

#> Adjusting the Data

#>   |                                                                              |======================================================================| 100%
#> 

re$Anderson_42_OriginalModel

#>  [1] 33625 71152 -6451 31540 33917 12625 66198  8276 15996 32449 39741 16490
#> [13] 20508 50502 49206 46410 29825  1620  3799  6588 17014 10863 22908  6514
#> [25]  7981 15288  3558 75895 31300 24716 24937