Bootstrap the AUC and conduct T-Tests for a collection of signatures.

Run bootstrapping of the AUC and derive the p-value for a 2-sample t-test for all signatures tested on a given dataset.

bootstrapAUC(
  SE_scored,
  annotationColName,
  signatureColNames,
  num.boot = 100,
  pb.show = TRUE
)

Arguments

SE_scored
a SummarizedExperiment object with genes as the row features and signature scores in the colData. There should also be a column of annotation data. Required.

annotationColName
a character string giving the column name in colData that contains the annotation data. Required.

signatureColNames
a vector of column names in the colData that contain the signature score data. Required.

num.boot
integer. The number of times to bootstrap the data. The default is 100.

pb.show
logical for whether to show a progress bar while running code. The default is TRUE.

Value

A list of length 3 returning a vector of p-values for a 2-sample t-test, bootstrapped AUC values, and an AUC value for using all scored values for all signatures specified in signatureColNames.

Examples

 # Run signature profiling
 choose_sigs <- TBsignatures[1]
 prof_indian <- runTBsigProfiler(TB_indian, useAssay = "logcounts",
                                 algorithm = "ssGSEA",
                                 combineSigAndAlgorithm = TRUE,
                                 signatures = choose_sigs,
                                 parallel.sz = 1)

#> Running ssGSEA

#> Warning: 1 genes with constant expression values throuhgout the samples.

#> Estimating ssGSEA scores for 1 gene sets.
#>   |                                                                              |                                                                      |   0%Using parallel with 1 cores
#>   |                                                                              |==                                                                    |   2%  |                                                                              |===                                                                   |   5%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   9%  |                                                                              |========                                                              |  11%  |                                                                              |==========                                                            |  14%  |                                                                              |===========                                                           |  16%  |                                                                              |=============                                                         |  18%  |                                                                              |==============                                                        |  20%  |                                                                              |================                                                      |  23%  |                                                                              |==================                                                    |  25%  |                                                                              |===================                                                   |  27%  |                                                                              |=====================                                                 |  30%  |                                                                              |======================                                                |  32%  |                                                                              |========================                                              |  34%  |                                                                              |=========================                                             |  36%  |                                                                              |===========================                                           |  39%  |                                                                              |=============================                                         |  41%  |                                                                              |==============================                                        |  43%  |                                                                              |================================                                      |  45%  |                                                                              |=================================                                     |  48%  |                                                                              |===================================                                   |  50%  |                                                                              |=====================================                                 |  52%  |                                                                              |======================================                                |  55%  |                                                                              |========================================                              |  57%  |                                                                              |=========================================                             |  59%  |                                                                              |===========================================                           |  61%  |                                                                              |=============================================                         |  64%  |                                                                              |==============================================                        |  66%  |                                                                              |================================================                      |  68%  |                                                                              |=================================================                     |  70%  |                                                                              |===================================================                   |  73%  |                                                                              |====================================================                  |  75%  |                                                                              |======================================================                |  77%  |                                                                              |========================================================              |  80%  |                                                                              |=========================================================             |  82%  |                                                                              |===========================================================           |  84%  |                                                                              |============================================================          |  86%  |                                                                              |==============================================================        |  89%  |                                                                              |================================================================      |  91%  |                                                                              |=================================================================     |  93%  |                                                                              |===================================================================   |  95%  |                                                                              |====================================================================  |  98%  |                                                                              |======================================================================| 100%

 # Bootstrapping
 booted <- bootstrapAUC(SE_scored = prof_indian, annotationColName = "label",
                        signatureColNames = names(choose_sigs), num.boot = 5)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

 booted

#> $`P-values`
#> [1] 0.0001587202
#> 
#> $`Boot AUC Values`
#>       tmp_aucs
#> [1,] 0.8261905
#> [2,] 0.7940447
#> [3,] 0.8281250
#> [4,] 0.8654244
#> [5,] 0.8627451
#> 
#> $`Non-Boot AUC Values`
#> [1] 0.8105263
#>