A function that implements the original methods for multiple TB signatures.

This function computes prediction foot multiple TB signatures based on their training models/methods. To avoid naming issues, the gene names for both training data and input gene sets have been updated using the checkGeneSymbols. TB signatures with available original models are: Anderson_42, Anderson_OD_51, Kaforou_27, Kaforou_OD_44, Kaforou_OD_53, Sweeney_OD_3, Maertzdorf_4, Verhagen_10, Jacobsen_3, Sambarey_HIV_10, Leong_24, Berry_OD_86, Berry_393, Bloom_OD_144, Suliman_RISK_4, Zak_RISK_16, Leong_RISK_29, and Zhao_NANO_6. The predicted score for each signature has been stored in the column data section of the input SummarizedExperiment study.

evaluateOriginalModel(
  input,
  geneSignaturesName,
  useAssay = 1,
  adj = 0.001,
  BPPARAM = BiocParallel::SerialParam(progressbar = TRUE)
)

Arguments

input
A SummarizedExperiment object with gene symbols as the assay row names.

geneSignaturesName
A character string/vector specifying the signature of interest. If any(geneSignaturesName == "") == TRUE run all available gene signatures' original models.

useAssay
A character string or an integer specifying the assay in the input. Default is the first assay in the assay list. Used for the test SummarizedExperiment object. Default is 1, indicating the first assay in the input.

adj
A small positive real number used in ComBat to solve for genes with 0 counts(rare cases). Default is 1e-3.

BPPARAM
An instance inherited from bplapply.

Value

A SummarizedExperiment object with predicted scores for each sample obtained from the signature's original model.

Examples

data("TB_hiv", package = "TBSignatureProfiler")
re <- evaluateOriginalModel(input = TB_hiv,
geneSignaturesName = c("Anderson_42", "Maertzdorf_15"), useAssay = "counts")

#>   |                                                                              |                                                                      |   0%

#> Now Evaluating: Anderson_42

#>   |                                                                              |======================================================================| 100%
#> 
#>   |                                                                              |                                                                      |   0%

#> Now Evaluating: Maertzdorf_15

#> Using batch =GSE_trainas a reference batch (this batch won't change)

#> Found2batches

#> Adjusting for0covariate(s) or covariate level(s)

#> Standardizing Data across genes

#> Fitting L/S model and finding priors

#> Finding parametric adjustments

#> Adjusting the Data

#>   |                                                                              |======================================================================| 100%
#> 

re$Anderson_42_OriginalModel

#>  [1] 29875 67665 -7104 30050 33128 12041 61535  5876 13917 30240 37941 15100
#> [13] 19287 49450 46805 43594 25553   941  1041  5566  9393  8363 11765  5411
#> [25]  4996 12948  1991 73154 29716 22816 21147