Create a table of results for t-tests and bootstrapped AUCs for multiple scored signatures.

This function collects the results of bootstrapping and t-tests for a scored gene expression dataset and presents them using a JavaScript table with an R interface, or as a data.frame.

tableAUC(
  SE_scored,
  annotationColName,
  signatureColNames,
  num.boot = 100,
  pb.show = TRUE,
  output = "DataTable",
  pROC = TRUE
)

Arguments

SE_scored
a SummarizedExperiment object with genes as the row features and signature scores in the colData. There should also be a column of annotation data. Required.

annotationColName
a character string giving the column name in colData that contains the annotation data. Required.

signatureColNames
a vector of column names in the colData that contain the signature score data. Required.

num.boot
integer. The number of times to bootstrap the data. The default is 100.

pb.show
logical for whether to show a progress bar while running code. The default is TRUE.

output
a character string indicating the table output format. Possible values are DataTable and data.frame. The default is DataTable.

pROC
logical. Should pROC AUC confidence intervals be used? Default is TRUE.

Value

A JavaScript table with an R interface using the DT package.

Examples

# Run signature profiling
 choose_sigs <- TBsignatures[c(1, 2)]
 prof_indian <- runTBsigProfiler(TB_indian, useAssay = "logcounts",
                                 algorithm = "ssGSEA",
                                 signatures = choose_sigs,
                                 parallel.sz = 1)

#> Running ssGSEA

#> Warning: 1 genes with constant expression values throuhgout the samples.

#> Estimating ssGSEA scores for 2 gene sets.
#>   |                                                                              |                                                                      |   0%  |                                                                              |==                                                                    |   2%  |                                                                              |===                                                                   |   5%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   9%  |                                                                              |========                                                              |  11%  |                                                                              |==========                                                            |  14%  |                                                                              |===========                                                           |  16%  |                                                                              |=============                                                         |  18%  |                                                                              |==============                                                        |  20%  |                                                                              |================                                                      |  23%  |                                                                              |==================                                                    |  25%  |                                                                              |===================                                                   |  27%  |                                                                              |=====================                                                 |  30%  |                                                                              |======================                                                |  32%  |                                                                              |========================                                              |  34%  |                                                                              |=========================                                             |  36%  |                                                                              |===========================                                           |  39%  |                                                                              |=============================                                         |  41%  |                                                                              |==============================                                        |  43%  |                                                                              |================================                                      |  45%  |                                                                              |=================================                                     |  48%  |                                                                              |===================================                                   |  50%  |                                                                              |=====================================                                 |  52%  |                                                                              |======================================                                |  55%  |                                                                              |========================================                              |  57%  |                                                                              |=========================================                             |  59%  |                                                                              |===========================================                           |  61%  |                                                                              |=============================================                         |  64%  |                                                                              |==============================================                        |  66%  |                                                                              |================================================                      |  68%  |                                                                              |=================================================                     |  70%  |                                                                              |===================================================                   |  73%  |                                                                              |====================================================                  |  75%  |                                                                              |======================================================                |  77%  |                                                                              |========================================================              |  80%  |                                                                              |=========================================================             |  82%  |                                                                              |===========================================================           |  84%  |                                                                              |============================================================          |  86%  |                                                                              |==============================================================        |  89%  |                                                                              |================================================================      |  91%  |                                                                              |=================================================================     |  93%  |                                                                              |===================================================================   |  95%  |                                                                              |====================================================================  |  98%  |                                                                              |======================================================================| 100%
#> 

 # Create table
 tableAUC(SE_scored = prof_indian, annotationColName = "label",
          signatureColNames = names(choose_sigs))

#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%

 # Create data.frame object
h <-  tableAUC(SE_scored = prof_indian, annotationColName = "label",
               signatureColNames = names(choose_sigs),
               output = "data.frame",
               num.boot = 5)

#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%

head(h)

#>        Signature P.value neg10xLog.P.value. LowerAUC    AUC UpperAUC
#> 1    Anderson_42   0e+00                Inf   0.9206 0.0000   0.9206
#> 2 Anderson_OD_51   0e+00                Inf   0.9955 0.0000   0.9955
#> 3    Anderson_42   2e-04            87.4837   0.9206 0.8105   0.9206
#> 4 Anderson_OD_51   0e+00           259.5843   0.9955 0.9832   0.9955